% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CELoss}
\alias{CELoss}
\title{Compute crossentropy loss}
\usage{
CELoss(gt_y)
}
\arguments{
\item{gt_y}{NumericMatrix. Element-wise multiplication of ground-truth label and softmax output}
}
\value{
Calculated loss.
}
\description{
Loss function
}
