% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{Compile}
\alias{Compile}
\title{Function to compile FullyLight model instance}
\usage{
Compile(
  loss = "categorical_crossentropy",
  optimizer = "adam",
  metric = "acc",
  learning_rate = 0.001,
  l1 = 0,
  l2 = 0,
  ...
)
}
\arguments{
\item{loss}{character string. Only support 'categorical_crossentropy', don't modify!}

\item{optimizer}{character string. 'adam' is default, also supports 'sgd'/'rmsprop'/'momentum'.}

\item{metric}{character string. Metric for classification, 'acc'/'recall'/'f1'/'precision', default is 'acc'.}

\item{learning_rate}{numeric. Learning rate, default is 1e-3.}

\item{l1}{numeric. L1 regularization penalty.}

\item{l2}{numeric. L2 regularization penalty.}

\item{...}{All other arguments related to the function.}
}
\value{
A list of compiled functions, including optimizer, loss function and metrics.
}
\description{
Function will call engine of all inside functions to compile FullyLight model,
one may put return into \code{FullyLight::Fit} function.
}
\examples{
compile_funcs <- Compile(loss = 'categorical_crossentropy',
                         optimizer = 'adam',
                         metric = 'f1',
                         learning_rate = 1e-3,
                         l1 = 0,
                         l2 = 0)
}
\author{
Kunhong Yu
}
