% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{Evaluate}
\alias{Evaluate}
\title{Function to evaluate FullyLight model instance}
\usage{
Evaluate(
  x,
  y,
  models,
  params,
  loss = "categorical_crossentropy",
  metric = "acc"
)
}
\arguments{
\item{x}{matrix. Input design matrix, shape: m, input_dim.}

\item{y}{matrix. One-hot form of labels, shape: m, num_classes.}

\item{models}{'Model' instance. If we have return from \code{FullyLight::Fit} as \code{res}, this parameter should be \code{res$models}.}

\item{params}{list. If we have return from \code{FullyLight::Fit} as \code{res}, this parameter should be \code{res$params}.}

\item{loss}{character string. Only support 'categorical_crossentropy', don't modify!}

\item{metric}{character string. Metric for classification, 'acc'/'recall'/'f1'/'precision', default is 'acc'.}
}
\value{
A list of results, including \code{eval_res} as metric on test set, \code{loss} for loss on test set.
}
\description{
Function will call engine of all inside functions to evaluate FullyLight model.
}
\examples{
x <- matrix(c(3, 7, 9, 10, 0, 9), nrow = 2)
y <- matrix(c(0, 1, 0, 1, 0, 0), nrow = 2, byrow = TRUE)
models <- FullyLight::Model(dims = c(3, 10, 3),
                            input_shape = 3,
                            hidden_activation = 'relu',
                            out_activation = 'softmax',
                            kernel_initializer = 'random',
                            l1 = 0,
                            l2 = 0)
compile_funcs <- FullyLight::Compile(loss = 'categorical_crossentropy',
                                     optimizer = 'adam',
                                     metric = 'f1',
                                     learning_rate = 1e-3,
                                     l1 = 0,
                                     l2 = 0)
res <- FullyLight::Fit(x = x,
                       y = y,
                       models = models,
                       compile_funcs = compile_funcs,
                       validation_rate = 0.01,
                       validation_data = NULL,
                       epochs = 1,
                       batch_size = 32,
                       verbose = 1,
                       shuffle = TRUE,
                       isimage = FALSE,
                       data_name = 'iris')
eval_res <- Evaluate(x = x, y = y, models = res$models, params = res$params,
                     loss = 'categorical_crossentropy', metric = 'f1')
}
\author{
Kunhong Yu
}
