% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{class}
\name{FullyModel-class}
\alias{FullyModel-class}
\title{Class S4 representing a FullyLight Model.}
\description{
Class defines an object with FullyLight Model's characteristics.
}
\section{Slots}{

\describe{
\item{\code{dims}}{Dimensions of NN}

\item{\code{input_shape}}{Input's shape}

\item{\code{hidden_activation}}{Hidden layer's activation}

\item{\code{out_activation}}{Output activation}

\item{\code{kernel_initializer}}{Kernel's initialization}

\item{\code{l1}}{L1 regularization}

\item{\code{l2}}{L2 regularization}
}}

