% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{Predict}
\alias{Predict}
\title{Function to make a prediction with trained FullyLight model}
\usage{
Predict(x, models, params, type = "probs")
}
\arguments{
\item{x}{matrix. Input design matrix, shape: m, input_dim.}

\item{models}{'Model' instance. If we have return from \code{FullyLight::Fit} as \code{res}, this parameter should be \code{res$models}.}

\item{params}{list. If we have return from \code{FullyLight::Fit} as \code{res}, this parameter should be \code{res$params}.}

\item{type}{character string. 'probs' or 'class', default is 'probs', in line with simple R \code{predict}.}
}
\value{
Predictions, shape: m, .
}
\description{
Function will call engine of all inside functions to make a prediction using pre-trained FullyLight model.
}
\examples{
x <- matrix(c(3, 7, 9, 10, 0, 9), nrow = 2)
y <- matrix(c(0, 1, 0, 1, 0, 0), nrow = 2, byrow = TRUE)
models <- FullyLight::Model(dims = c(3, 10, 3),
                            input_shape = 3,
                            hidden_activation = 'relu',
                            out_activation = 'softmax',
                            kernel_initializer = 'random',
                            l1 = 0,
                            l2 = 0)
compile_funcs <- FullyLight::Compile(loss = 'categorical_crossentropy',
                                     optimizer = 'adam',
                                     metric = 'f1',
                                     learning_rate = 1e-3,
                                     l1 = 0,
                                     l2 = 0)
res <- FullyLight::Fit(x = x,
                       y = y,
                       models = models,
                       compile_funcs = compile_funcs,
                       validation_rate = 0.01,
                       validation_data = NULL,
                       epochs = 1,
                       batch_size = 32,
                       verbose = 1,
                       shuffle = TRUE,
                       isimage = FALSE,
                       data_name = 'iris')
preds <- Predict(x = x, models = res$models, params = res$params, type = 'probs')
}
\author{
Kunhong Yu
}
