% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{Summary}
\alias{Summary}
\title{Summary of FullyLight model}
\usage{
Summary(models)
}
\arguments{
\item{models}{'Model' instance, from \code{FullyLight::Model} function.}
}
\description{
Keras-like summary of the model
}
